+++ src/app/programs/forge/page.tsx
@@ -30,6 +30,7 @@ const Page = () => {
         textColor="text-white"
         backgroundColor="bg-acm-gray-200"
         image={Forge1}
+        name={"forge"}
       />
       <WhatWeDo />
       <FAQ />
