+++ src/components/programs/joinus.tsx
@@ -4,11 +4,14 @@ import Image from "next/image";
 import { StaticImageData } from "next/image";
 import Link from "next/link";
 import { motion } from "framer-motion";
+import getCurrentApps from "@/utils/applications/currentapplications";
+import { getAppStatus } from "@/utils/applications/applicationstatus";
 
 interface ProgramProps {
   textColor: string;
   backgroundColor: string;
   image: StaticImageData;
+  name: string;
 }
 
 // Animation variants
@@ -34,7 +37,11 @@ const ImageAnimation = {
   transition: { duration: 0.6, delay: 0.3 },
 };
 
-const Joinus = ({ textColor, backgroundColor, image }: ProgramProps) => {
+const Joinus = ({ textColor, backgroundColor, image, name }: ProgramProps) => {
+  const { data, isLoading } = getCurrentApps();
+  const currentApps = data?.currentApps;
+  const { status, appLink } = getAppStatus(name, currentApps ?? []);
+
   return (
     <div className="mx-auto grid w-10/12 items-center md:grid-cols-2">
       <div className="flex flex-col">
@@ -47,12 +54,20 @@ const Joinus = ({ textColor, backgroundColor, image }: ProgramProps) => {
 
         <div className="flex justify-center py-4 md:py-0">
           <motion.div {...ButtonAnimation}>
-            <Link
-              href="/"
-              className={`rounded-lg p-2 text-2xl md:mt-0 md:px-10 md:py-4 md:text-4xl ${textColor} ${backgroundColor}`}
-            >
-              Apply Here!
-            </Link>
+            {appLink && appLink !== "/" ? (
+              <Link
+                href={appLink}
+                className={`rounded-lg p-2 text-2xl md:mt-0 md:px-10 md:py-4 md:text-4xl ${textColor} ${backgroundColor} ${appLink === "/" ? "cursor-not-allowed" : ""}`}
+              >
+                {isLoading ? "Loading..." : status} APPLICATION
+              </Link>
+            ) : (
+              <p
+                className={`rounded-lg p-2 text-2xl md:mt-0 md:px-10 md:py-4 md:text-4xl ${textColor} ${backgroundColor} ${appLink === "/" ? "cursor-not-allowed" : ""}`}
+              >
+                {isLoading ? "Loading..." : status}
+              </p>
+            )}
           </motion.div>
         </div>
       </div>
